[CHORE] #9 - 폴더링 수정
