+++ dnd-11th-4-iOS/dnd-11th-4-iOS/Presetation/Home/Model/ColorWithOpacityModel.swift
@@ -1,0 +1,12 @@
+//
+//  ColorWithOpacityModel.swift
+//  dnd-11th-4-iOS
+//
+//  Created by 황찬미 on 8/8/24.
+//
+
+struct ColorWithOpacityModel {
+    let name: String
+    let opacity: Int
+    let colorType: OpacityColorType
+}
